Component module: ComponentObject->BrowserObject, TypeKey->BrowserObjectId, ComponentRuntime->ComponentDriver (internal split+rename, behavior-preserving)
