+++ Alis.Reactive/PlanModel/BrowserObjectContracts.cs
@@ -10,7 +10,7 @@ namespace Alis.Reactive.PlanModel
         internal IReadOnlyDictionary<string, BrowserObjectContract> Snapshot() =>
             new Dictionary<string, BrowserObjectContract>(_byTypeKey);
 
-        internal BrowserObjectContract Require(TypeKey key)
+        internal BrowserObjectContract Require(BrowserObjectId key)
         {
             if (!_byTypeKey.TryGetValue(key.Value, out var objectContract))
                 throw new InvalidOperationException($"browser object contract '{key.Value}' is not registered in this plan.");
@@ -18,38 +18,38 @@ namespace Alis.Reactive.PlanModel
             return objectContract;
         }
 
-        internal void DeclareInputValueContract(TypeKey key, InputValueContract contract)
+        internal void DeclareInputValueContract(BrowserObjectId key, InputValueContract contract)
         {
             DeclareObject(key);
             contract.Enrich(Require(key));
         }
 
-        internal void DeclareObject(TypeKey key)
+        internal void DeclareObject(BrowserObjectId key)
         {
             if (!_byTypeKey.ContainsKey(key.Value))
                 _byTypeKey[key.Value] = new BrowserObjectContract();
         }
 
-        internal void DeclareProperty(TypeKey typeKey, ObjectPropertyContract contract) =>
+        internal void DeclareProperty(BrowserObjectId typeKey, ObjectPropertyContract contract) =>
             Require(typeKey).Declare(contract);
 
-        internal ObjectMethod DeclareMethod(TypeKey typeKey, ObjectMethodContract contract) =>
+        internal ObjectMethod DeclareMethod(BrowserObjectId typeKey, ObjectMethodContract contract) =>
             Require(typeKey).Declare(contract);
 
-        internal void DeclareEvent(TypeKey typeKey, ObjectEventContract contract) =>
+        internal void DeclareEvent(BrowserObjectId typeKey, ObjectEventContract contract) =>
             Require(typeKey).Declare(contract);
 
         internal void RegisterPlugin(PluginContract contract)
         {
-            if (_byTypeKey.ContainsKey(contract.TypeKey.Value))
+            if (_byTypeKey.ContainsKey(contract.BrowserObjectId.Value))
                 throw new InvalidOperationException($"Plugin '{contract.Name.Value}' is already registered.");
 
-            _byTypeKey[contract.TypeKey.Value] = contract.ToBrowserObjectContract();
+            _byTypeKey[contract.BrowserObjectId.Value] = contract.ToBrowserObjectContract();
         }
 
         internal MethodSignature DeclarePluginMethod(PluginMethodRequirement methodRead)
         {
-            var typeKey = TypeKey.Plugin(methodRead.PluginName);
+            var typeKey = BrowserObjectId.Plugin(methodRead.PluginName);
 
             if (!_byTypeKey.ContainsKey(typeKey.Value))
                 throw new InvalidOperationException(
@@ -65,7 +65,7 @@ namespace Alis.Reactive.PlanModel
 
         internal void DeclarePluginProperty(PluginPropertyRequirement propertyRead)
         {
-            var typeKey = TypeKey.Plugin(propertyRead.PluginName);
+            var typeKey = BrowserObjectId.Plugin(propertyRead.PluginName);
 
             if (!_byTypeKey.ContainsKey(typeKey.Value))
                 throw new InvalidOperationException(
