+++ Alis.Reactive/PlanModel/PlanDocument.cs
@@ -10,20 +10,20 @@ namespace Alis.Reactive.PlanModel
     {
         private readonly PlanIdentity _identity;
         private readonly IReadOnlyDictionary<string, BrowserObjectContract> _types;
-        private readonly IReadOnlyDictionary<string, ComponentObject> _components;
+        private readonly IReadOnlyDictionary<string, BrowserObject> _components;
         private readonly IReadOnlyList<Behavior> _behaviors;
 
         public int Version => 3;
         public string PlanId => _identity.PlanIdForJson;
         public PlanScope Scope => _identity.ScopeForJson;
         public IReadOnlyDictionary<string, BrowserObjectContract> Types => _types;
-        public IReadOnlyDictionary<string, ComponentObject> Components => _components;
+        public IReadOnlyDictionary<string, BrowserObject> Components => _components;
         public IReadOnlyList<Behavior> Behaviors => _behaviors;
 
         internal PlanDocument(
             PlanIdentity identity,
             IReadOnlyDictionary<string, BrowserObjectContract> types,
-            IReadOnlyDictionary<string, ComponentObject> components,
+            IReadOnlyDictionary<string, BrowserObject> components,
             IReadOnlyList<Behavior> behaviors)
         {
             _identity = identity;
